do the js shuffle 🕺
