+++ js/forum.js
@@ -1,2 +1,1 @@
-export * from './src/common';
 export * from './src/forum';
