+++ js/src/forum/extend/index.js
@@ -1,0 +1,9 @@
+import SettingsPage from "./SettingsPage";
+import Discussion from "./Discussion";
+import User from "./User";
+
+export default (app) => {
+    Discussion(app);
+    SettingsPage(app);
+    User(app);
+}
