+++ js/src/forum/pages/index.js
@@ -1,0 +1,7 @@
+import routes from "./routes";
+import PrivateDiscussionsPage from "./PrivateDiscussionsPage";
+
+export default (app) => {
+    routes(app);
+    PrivateDiscussionsPage(app);
+}
