+++ js/src/forum/pages/routes.js
@@ -1,0 +1,7 @@
+import IndexPage from 'flarum/components/IndexPage';
+import PrivateDiscussionsUserPage from "./PrivateDiscussionsUserPage";
+
+export default (app) => {
+    app.routes.byobuUserPrivate = { path: '/u/:username/private', component: PrivateDiscussionsUserPage };
+    app.routes.byobuPrivate = {path: '/private', component: IndexPage};
+}
