--- src/Listeners/GetModelIsPrivate.php
@@ -1,21 +1,0 @@
-<?php
-
-/*
- * This file is part of fof/byobu.
- *
- * Copyright (c) 2019 FriendsOfFlarum.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace FoF\Byobu\Listeners;
-
-use Flarum\Event\GetModelIsPrivate as Event;
-
-class GetModelIsPrivate
-{
-    public function handle(Event $event)
-    {
-    }
-}
